Fix assert for structs with no decorations
Previously, if a shader didn't decorate all the members (or
specifically, the last member) with any decorations/names,
an assert would trigger in ParseTypes because the OpTypeStruct's
member_count would be 0.  This fixes that by hardcoding member count for
Struct types to word_count - 2 (based on the binary encoding of
OpTypeStruct).