Add support for buffer pointers in buffers
A buffer pointer is a pointer to a struct in the physical address
space. The struct is decorated as is any buffer. These are currently
generated by glsang for the GLSL buffer_reference feature.

This adds support for buffer pointers in buffers. Buffer pointers
in buffers are treated as if the struct they point to is embedded
in the buffer the pointer appears in at the point the pointer
appears in. A new type flag bit is added to indicate a struct is
actually a reference/pointer.

This commit does not add support for buffer pointers outside of
buffers, for example, as local variables. That support is TBD.