Fix generating invalid YAML files (#282)
YAML 1.0 demands that the separator between a YAML directive and its
value is a colon. This wasn't the case for the YAML files generated by
spirv-reflect, as it used a space as separator instead. This resulted in
tools parsing YAML stricly not being able to process the YAML files
generated by spirv-reflect.

Newer YAML versions demand a space as separator between YAML directive
and its value instead.

To fix this, this commit increases the YAML version to 1.1. This is done
instead of replacing the separator to improve compatibility with
tooling, as many tools don't support YAML 1.0 and require YAML 1.1 as
minimal version.