Fix undefined enums
There are a few places where an enum value is purpoposly getting a value
that is not one of the enum values. This is causing a problem when
testing. I have a quick fix to make those values `int` instead of the
enum type.

The variables could contain `-1` as a special value to indicate that no
storage class or builtin is used. We cannot change the enum class to add
`-1` because it is defined in spirv headers.